Solution to Count Students Scoring Above a Threshold

+++ Count Students Scoring Above a Threshold@@67877d61dc107486ad5c2747/code.c
@@ -14,7 +14,7 @@ int main(){
         scanf("%s",s[i].name);
         scanf("%f",&s[i].marks);
     }
-    int k;
+    float k;
     int count=0;
     scanf("%f",&k);
     for(int i=0;i<n;i++){
